oracle, leveldb, cmd : move midpointPrice to the leveldb driver package and modify the way initializing bytes array
